+++ jee-lab-core/pom.xml
@@ -4,9 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>jee-lab</groupId>
     <artifactId>jee-lab-core</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>${jee.lab.version}</version>
     <parent>
         <groupId>jee-lab</groupId>
         <artifactId>jee-lab</artifactId>
@@ -19,12 +18,12 @@
         <dependency>
             <groupId>com.google.code.findbugs</groupId>
             <artifactId>annotations</artifactId>
-            <version>2.0.1</version>
+            <version>${google.code.findbugs.version}</version>
         </dependency>
         <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
-            <version>23.0</version>
+            <version>${google.guava.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
