+++ jee-lab-core/src/main/java/com/rozzer/common/Named.java
@@ -7,7 +7,6 @@ public interface Named {
     /**
      * @return name for object
      */
-    @Nonnull
     String getName();
 
     /**
