+++ jee-lab-db/pom.xml
@@ -4,9 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>jee-lab</groupId>
     <artifactId>jee-lab-db</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>${jee.lab.version}</version>
+    <parent>
+        <groupId>jee-lab</groupId>
+        <artifactId>jee-lab</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
     <build>
         <finalName>jee-lab-db</finalName>
         <plugins>
@@ -20,16 +24,11 @@
             </plugin>
         </plugins>
     </build>
-    <parent>
-        <groupId>jee-lab</groupId>
-        <artifactId>jee-lab</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
     <dependencies>
         <dependency>
             <groupId>jee-lab</groupId>
             <artifactId>jee-lab-core</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>${jee.lab.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
@@ -37,6 +36,7 @@
             <artifactId>spring-boot-starter-data-jpa</artifactId>
             <version>${spring.boot.version}</version>
         </dependency>
+        <!--TODO: h2 DB is WA. Please use PostgresSQL or some real DB-->
         <dependency>
             <groupId>com.h2database</groupId>
             <artifactId>h2</artifactId>
