+++ jee-lab-security/pom.xml
@@ -2,29 +2,28 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>jee-lab-controller</artifactId>
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>jee-lab-security</artifactId>
     <version>${jee.lab.version}</version>
     <parent>
-        <groupId>jee-lab</groupId>
         <artifactId>jee-lab</artifactId>
+        <groupId>jee-lab</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <build>
-        <finalName>jee-lab-controller</finalName>
+        <finalName>jee-lab-security</finalName>
     </build>
     <dependencies>
         <dependency>
-            <groupId>jee-lab</groupId>
-            <artifactId>jee-lab-core</artifactId>
-            <version>${jee.lab.version}</version>
-            <scope>compile</scope>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-security</artifactId>
+            <version>${spring.boot.version}</version>
         </dependency>
         <dependency>
-            <groupId>jee-lab</groupId>
-            <artifactId>jee-lab-core</artifactId>
-            <version>${jee.lab.version}</version>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-thymeleaf</artifactId>
+            <version>${spring.boot.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -32,20 +31,11 @@
             <version>${spring.boot.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <version>${spring.boot.version}</version>
+            <groupId>org.springframework.security</groupId>
+            <artifactId>spring-security-test</artifactId>
+            <version>${spring.version}</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>3.0.1</version>
-            <scope>provided</scope>
-        </dependency>
     </dependencies>
-
-
-
 
 </project>
