+++ jee-lab-security/src/main/java/com/rozzer/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.rozzer.config;
+package com.rozzer.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
