+++ jee-lab-ui/pom.xml
@@ -6,8 +6,8 @@
   <modelVersion>4.0.0</modelVersion>
 
   <artifactId>jee-lab-ui</artifactId>
-  <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
+  <version>${jee.lab.version}</version>
   <parent>
     <groupId>jee-lab</groupId>
     <artifactId>jee-lab</artifactId>
