+++ pom.xml
@@ -5,7 +5,10 @@
     <artifactId>jee-lab</artifactId>
     <version>1.0-SNAPSHOT</version>
     <properties>
+        <jee.lab.version>1.0-SNAPSHOT</jee.lab.version>
         <spring.boot.version>2.0.5.RELEASE</spring.boot.version>
+        <google.guava.version>23.0</google.guava.version>
+        <google.code.findbugs.version>2.0.1</google.code.findbugs.version>
         <spring.version>5.0.5.RELEASE</spring.version>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
@@ -17,6 +20,7 @@
 
 
     <modules>
+        <module>jee-lab-security</module>
         <module>jee-lab-aggregator</module>
         <module>jee-lab-controller</module>
         <module>jee-lab-core</module>
